async helpers in AndroidAsyncTask (#3631)
* [Xamarin.Android.Build.Tasks] async helpers in AndroidAsyncTask

Right now there is a lot of boilerplate when using `AsyncTask`:

    public override bool RunTask ()
    {
        Yield ();
        try {
            var task = this.RunTask (DoExecute);

            task.ContinueWith (Complete);

            base.RunTask ();
        } finally {
            Reacquire ();
        }

        return !Log.HasLoggedErrors;
    }

We don't want to remember all that... or have to consistently get it
right.

Then if you want to use `Parallel.ForEach` it gets even more complex:

    this.RunTask (
        () => this.ParallelForEach (SomeArray, DoExecute)
    ).ContinueWith (Complete);

In d4c8f077, we created a base `AndroidAsyncTask` for error handling.
Let's put this weird `Yield-try-finally-Reacquire` logic in there,
too!

So instead, you would override:

    public override System.Threading.Tasks.Task RunTaskAsync ()

This method is:

* Already on a background thread.
* You can use `async` / `await` nicely.

I also added helper methods, so that instead of:

    this.ParallelForEach (ManifestFiles, ProcessManifest);

You can instead do:

    await this.WhenAll (ManifestFiles, ProcessManifest);

This plays well with `async` / `await`.

In refactoring this, I found some places where `<ResolveAssemblies/>`
was accessing `Log.Log*` on a background thread... I fixed that.

* Docs for Android*Task base classes

* Update MSBuildBestPractices.md